Added popups to better user experience

+++ lib/pages/view_file_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:wetra_app/custom_classes/api_const.dart';
+import 'package:wetra_app/custom_classes/login_register_popup.dart';
 import 'package:wetra_app/custom_classes/uploaded_file.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:wetra_app/custom_classes/user.dart';
@@ -35,6 +36,13 @@ class _ViewFileScreenState extends State<ViewFileScreen> {
         'Authorization': 'Bearer $token',
       },
     );
+
+    if (response.statusCode != 200) {
+      // OtherPopups.createPopup(context, 'Error', 'There was an error deleting the file please try again later');
+      throw Exception('Failed to delete file');
+    } else {
+      // OtherPopups.createPopup(context, 'File Deleted', 'The file has been deleted successfully');
+    }
 
     Navigator.of(context).pop();
   }
@@ -81,7 +89,28 @@ class _ViewFileScreenState extends State<ViewFileScreen> {
                       ),
                       child: const Text('Delete'),
                       onPressed: () {
-                        _deleteFile();
+                        showDialog<String>(
+                          context: context,
+                          builder: (BuildContext context) => AlertDialog(
+                              title: const Text('DELETE FILE'),
+                              content: const Text(
+                                  'Are you sure you want to delete this file. This action cannot be undone!'),
+                              actions: <Widget>[
+                                TextButton(
+                                  onPressed: () =>
+                                      Navigator.pop(context, 'CANCEL'),
+                                  child: const Text('CANCEL'),
+                                ),
+                                TextButton(
+                                  onPressed: () {
+                                    _deleteFile();
+                                    Navigator.pop(context, 'OK');
+                                    Navigator.of(context).pop();
+                                  },
+                                  child: const Text('OK'),
+                                ),
+                              ]),
+                        );
                       })),
               visible: User.getUser().user.isAdmin == 0 ? false : true,
             ),
